+++ release/components/body/body-row.component.js
@@ -123,6 +123,7 @@ var DataTableBodyRowComponent = (function () {
     DataTableBodyRowComponent.decorators = [
         { type: core_1.Component, args: [{
                     selector: 'datatable-body-row',
+                    changeDetection: core_1.ChangeDetectionStrategy.OnPush,
                     template: "\n    <div\n      *ngFor=\"let colGroup of columnsByPin; let i = index; trackBy: trackByGroups\"\n      class=\"datatable-row-{{colGroup.type}} datatable-row-group\"\n      [ngStyle]=\"stylesByGroup(colGroup.type)\">\n      <datatable-body-cell\n        *ngFor=\"let column of colGroup.columns; let ii = index; trackBy: columnTrackingFn\"\n        tabindex=\"-1\"\n        [row]=\"row\"\n        [expanded]=\"expanded\"\n        [isSelected]=\"isSelected\"\n        [rowIndex]=\"rowIndex\"\n        [column]=\"column\"\n        [rowHeight]=\"rowHeight\"\n        (activate)=\"onActivate($event, ii)\">\n      </datatable-body-cell>\n    </div>\n  "
                 },] },
     ];