+++ docs/changelog.md
@@ -1,4 +1,9 @@
 # Changelog
+## 10.0.4
+- Perf: Changed everything internal to OnPush
+- Perf: Fixed header cell outlet cd
+- Bug: Fixed sort mutating original rows
+
 ## 10.0.3
 - Perf: Better performance in cells